adjust all timestamps by the max range offset, fixes zoom/rendering bug for big start timestamp files

+++ src/platform.odin
@@ -51,12 +51,12 @@ mouse_up :: proc "contextless" (x, y: f32) {
 }
 
 @export
-scroll :: proc "contextless" (x, y: f32) {
+scroll :: proc "contextless" (x, y: f64) {
 	zoom_velocity += y
 }
 
 @export
-zoom :: proc "contextless" (x, y: f32) {
+zoom :: proc "contextless" (x, y: f64) {
 	zoom_velocity += y
 }
